dom/tests/mochitest/pointerlock/file_pointerLockCSSDisplay.html

<!DOCTYPE HTML>
<html>
  <!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=633602
  -->
  <head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
    </script>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script type="application/javascript" src="mouselock_util.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <style>
    </style>
  </head>
  <body onload="start();">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602
    </a>
    <div id="div">
    </div>
    <pre id="test">
      <script type="text/javascript">
        /*
         * Test for Bug 633602
         * Test to check that pointer cannot be locked if the locking element
         * have its CSS display property set to none.
         */
        SimpleTest.waitForExplicitFinish();
        var pointer = navigator.mozPointer;
        var div = document.getElementById("div");

        document.addEventListener("mozfullscreenchange", function (e) {
          if (document.mozFullScreen && document.mozFullScreenElement === div) {
            // If the calling element have a style of "display: none", then
            // pointer should not lock and call the failure callback.
            // If element does not have the said style, then lock is enabled.
            // First lock case should cause success callback as default display
            // is set to "block"
            pointer.lock(div, function() {
              // Failure callback - should fire when element is attached to
              // a DOM Tree and in fullscreen mode
              ok(pointer.isLocked, "Mouse locked as display is \"block\"");
              div.style.display = "none";
              pointer.lock (div, function() {
                is(pointer.isLocked, false, "Mouse should not be locked")
                document.mozCancelFullScreen();
              }, function() {
                is(pointer.isLocked, false, "Mouse failed to lock as " +
                  "display is set to \"none\"");
              });
              document.mozCancelFullScreen();
            }, function () {
              // Success Callback
              // Test failed
              ok(pointer.isLocked, "Mouse locked when style display " +
                "is set to \"block\"");
              document.mozCancelFullScreen();
            });
          } else {
            SimpleTest.finish();
          }
        }, false);

        function start()
        {
          SimpleTest.waitForFocus(function() {
            // Begin test with no style element set;
            // Defaults to display: block
            div.mozRequestFullScreen();
          });
        }
      </script>
    </pre>
  </body>
</html>
